podman based testexecution via testctl
We use a podman pod instead of docker-compose to run the tests for
better isolation, and for less interaction with the an existing
environment. This should also make integration into a CI system easier.

Some adjustmenets are required because everything connects over
localhost in a pod.

This patch also adds:
* A mechanism to apply an overlay during startup (used to apply the
  test overlay config.demo)
* Fixed the roundcube update from source code

+++ docker/webapp/init.sh
@@ -4,6 +4,7 @@ set -x
 
 cd /opt/app-root/src/
 
+REBUILD=false
 # Update the sourcecode if available.
 # This also copies the .env files that is required if we don't provide
 # a configuration via the environment.
@@ -22,6 +23,34 @@ if [ -d /src/kolabsrc.orig ]; then
         --exclude=.gitignore \
         /src/kolabsrc.orig/ /opt/app-root/src/ | tee /tmp/rsync.output
 
+    REBUILD=true
+fi
+
+if [ -d /src/overlay ]; then
+    echo "----> Applying overlay"
+    rsync -av \
+        --exclude=vendor \
+        --exclude=composer.lock \
+        --exclude=node_modules \
+        --exclude=package-lock.json \
+        --exclude=public \
+        --exclude=storage \
+        --exclude=resources/build \
+        --exclude=bootstrap \
+        --exclude=.gitignore \
+        /src/overlay/ /opt/app-root/src/ | tee /tmp/rsync-overlay.output
+
+    REBUILD=true
+fi
+
+# If we want to rely on the environment for configuration
+if [ $NOENVFILE ]; then
+    echo "----> removing envfile"
+    rm -f .env
+fi
+
+
+if [ $REBUILD ]; then
     rm -rf storage/framework
     mkdir -p storage/framework/{sessions,views,cache}
